+++ src/command_modules/azure-cli-sf/HISTORY.rst
@@ -4,11 +4,14 @@ Release History
 ===============
 
 unreleased
-+++++++++++++++
+++++++++++
+* Deprecating all commands in favor of Service Fabric CLI (sfctl)
+
+1.0.7 (2017-08-28)
+++++++++++++++++++
 * Simplified registry user/pass rules for command.
 * Fixed password prompt for user even after passing in the param.
 * Supports None registry_cred.
-* Deprecating all commands in favor of Service Fabric CLI (sfctl)
 
 1.0.6 (2017-08-11)
 ++++++++++++++++++