+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -3,8 +3,9 @@
 Release History
 ===============
 
-unreleased
-++++++++++
+2.0.13 (2017-08-28)
++++++++++++++++++++
+* Enable set blob tier
 * `storage account create/update`: Add `--bypass` and `--default-action` arguments to support service tunneling.
 * `storage account network-rule`: Added commands to add VNET rules and IP based rules.
 * Enable service encryption by customer managed key